Reordered tests and added basic XML response body handling to BoundConnection
If the BoundConnection gets a response with content-type of application/xml,
it will use xml.dom.minidom to parse the response body accordingly and return
the document's documentElement for the body portion of the result.

Additionally, if the status code is 400+, we dispatch to the exceptions.throw()
function after extracting message and details from the response body (if they
exist).

This means that the established rackspace cloud fault conditions should all map
to the exception classes.