Cheat on timestamp handling and HTTP connection management
The httplib.HTTP(s)Connection classes are persnickety about
state between requests and it makes things somewhat difficult
to manage.  So, as an initial step, we just close and reopen
the connection per request, which is heavy-handed but effective.

Since native Python timezone handling is rather uninspiring,
backed out the conversion of timestamp strings to datetime.datetime
objects, as this isn't worth fighting in the near term.  Later, though.