Account class composing a cloud API provider with auth and session mgmt.
The thunderhead.Account class allows specification of:
* a cloud API provider (initially just thunderhead.rackspace)
* access credentials idiomatic to that provider

From there, the resulting account object has a session based on the
provider API, and includes all management methods offered by the
provider, wrapped so as to be bound to the underlying session/connection.

This is largely a matter of convenience, but the little things in life
are what make all the difference.